requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
babel==2.14.0
    # via
    #   -r requirements.in
    #   flask-babel
blinker==1.7.0
    # via flask
certifi==2023.11.17
    # via requests
cffi==1.16.0
    # via
    #   cryptography
    #   pynacl
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via flask
cryptography==41.0.7
    # via pyjwt
deprecated==1.2.14
    # via pygithub
flask==3.0.1
    # via
    #   -r requirements.in
    #   flask-babel
flask-babel==4.0.0
    # via -r requirements.in
gunicorn==21.2.0
    # via -r requirements.in
idna==3.6
    # via requests
itsdangerous==2.1.2
    # via flask
jinja2==3.1.3
    # via
    #   -r requirements.in
    #   flask
    #   flask-babel
markupsafe==2.1.3
    # via
    #   jinja2
    #   werkzeug
packaging==23.2
    # via gunicorn
pycparser==2.21
    # via cffi
pygithub==2.1.1
    # via -r requirements.in
pyjwt[crypto]==2.8.0
    # via
    #   pygithub
    #   pyjwt
pynacl==1.5.0
    # via pygithub
python-dateutil==2.8.2
    # via pygithub
pytz==2023.3.post1
    # via flask-babel
requests==2.31.0
    # via
    #   -r requirements.in
    #   pygithub
six==1.16.0
    # via python-dateutil
typing-extensions==4.9.0
    # via pygithub
urllib3==2.1.0
    # via
    #   pygithub
    #   requests
werkzeug==3.0.1
    # via flask
wrapt==1.16.0
    # via deprecated
